requirements.txt
# AGCH-Impl Requirements
# Core dependencies for the project

# PyTorch and Lightning
torch>=2.9.0
lightning>=2.6.0
torchmetrics>=1.0.0

# Hydra configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Data handling
h5py>=3.10.0
numpy>=1.24.0

# Utilities
rootutils>=1.0.0
rich>=13.0.0
python-dotenv>=1.0.0

# Development tools
black>=24.0.0
flake8>=7.0.0
pytest>=8.0.0
pre-commit>=3.0.0
